requirements.txt
Pillow>=10.2.0
pyzbar>=0.1.9
qrcode>=7.4.2
flowlauncher>=0.1.0